feat(S13P21A302-161): 인플루언서 회원가입 구현

+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/content/dto/response/ContentDetailResponse.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public record ContentDetailResponse(
-    Long contentId,
+    Integer contentId,
     String contentType,
     String title,
     String description,
