+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/insight/dto/response/dailyAccountInsight/MonthlyAccountInsightResponse.java
@@ -4,24 +4,24 @@ import java.math.BigInteger;
 import java.time.YearMonth;
 
 public record MonthlyAccountInsightResponse(
-    Long platformAccountId,
-    String platformTypeId,
-    BigInteger totalViews,
-    Integer totalFollowers,
-    Integer totalContents,
-    BigInteger totalLikes,
-    BigInteger totalComments,
-    YearMonth snapshotDate
+        Integer platformAccountId,
+        String platformTypeId,
+        BigInteger totalViews,
+        Integer totalFollowers,
+        Integer totalContents,
+        BigInteger totalLikes,
+        BigInteger totalComments,
+        YearMonth snapshotDate
 ) {
   public static MonthlyAccountInsightResponse of(
-      Long platformAccountId,
-      String platformTypeId,
-      BigInteger totalViews,
-      Integer totalFollowers,
-      Integer totalContents,
-      BigInteger totalLikes,
-      BigInteger totalComments,
-      YearMonth snapshotDate
+          Integer platformAccountId,
+          String platformTypeId,
+          BigInteger totalViews,
+          Integer totalFollowers,
+          Integer totalContents,
+          BigInteger totalLikes,
+          BigInteger totalComments,
+          YearMonth snapshotDate
   ) {
     return new MonthlyAccountInsightResponse(
         platformAccountId,
