+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/insight/dto/response/dailyAccountInsight/MonthlyAccountViewsResponse.java
@@ -4,15 +4,15 @@ import java.math.BigInteger;
 import java.time.YearMonth;
 
 public record MonthlyAccountViewsResponse(
-    Long platformAccountId,
-    String platformTypeId,
-    BigInteger totalViews,
-    YearMonth snapshotDate
+        Integer platformAccountId,
+        String platformTypeId,
+        BigInteger totalViews,
+        YearMonth snapshotDate
 ) { public static MonthlyAccountViewsResponse of(
-    Long platformAccountId,
-    String platformTypeId,
-    BigInteger totalViews,
-    YearMonth snapshotDate
+        Integer platformAccountId,
+        String platformTypeId,
+        BigInteger totalViews,
+        YearMonth snapshotDate
 ) {
   return new MonthlyAccountViewsResponse(platformAccountId, platformTypeId, totalViews, snapshotDate);
 }}
