+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/insight/dto/response/dailyTrendingInsight/TrendingInfluencerResponse.java
@@ -3,10 +3,10 @@ package com.ssafy.leaper.domain.insight.dto.response.dailyTrendingInsight;
 import com.ssafy.leaper.domain.insight.entity.DailyTrendingInfluencer;
 
 public record TrendingInfluencerResponse(
-    Long influencerId,
-    Integer influencerRank,
-    String accountNickname,
-    Integer totalFollowers
+        Integer influencerId,
+        Integer influencerRank,
+        String accountNickname,
+        Integer totalFollowers
 ) {
   public static TrendingInfluencerResponse of(
       DailyTrendingInfluencer dpi,
